Adding ability to add and remove sites from the team update form.
In the past I had been using jquery to update the chosen selections
because I hadn't figureged out how to do it from the controller.
Example of how to set selections in Chosen() is available here.

Followed the same pattern to add te sites form as I had used before
to add the resources form. This strikes me as not a great method
because it adds multiple csrf keys to the page. Can't yet find a
better way to get a single form to display accross multiple on a
page.