Implement as EntityInterface to be compatible with teamAuthority check.
EntityInterfaces must implement a getId() method, and because this
has a composite primary key, I didn't think it was appropriatee.
The return might need to be massaged a bit, or the way id are
used might need to be tweeked to make sure opperations that use
this method perform as expected.