feat: automatically add global admin to bypass filter on upgrade
Before v4.1 global admins were the only user who could use the
bypass teams filter feature. In 4.1, that feature can optionally
be tuned to accomodate any level of user role in the module config.
As such, to keep behavior consistent when upgrading, we need to add
the global admin to the settting.