fix: remove permissions unless granted by team
Acl exhibits some counterintuitive behavior, inheritance, etc.
You can't seem to deny all then allow where team grants permit
becuase of hard to explain inheritance rules. So, resorting to
denying unless the team permission resolves to true--which is very
awkward in the acl syntax because you have to do a double-negative:
(deny when when permission assertion is not true)